Verify existance of provided sketch file & serve its directory
If the --sketch option is specified (#4), the provided sketch path is
checked to verify that it is an extant .js file. If it passes those
checks, the directory in which the file is located is served as the
server root.

This sets us up to render an index.html page that references the
sketch file, as well as the p5 files served within /lib.