Construct an HTML page to serve the provided sketch file
This enables --sketch mode, and permits the user to pass a .js file
name instead of a directory path: the CLI will detect the script and
switch into --sketch mode even if that argument is not specified. This
should lower the barrier to entry, by making `p5 run mysketch.js` a
valid command!

Once sketch mode is enabled, we

- Serve the directory in which the specified sketch is found,
- Serve p5 from within the CLI itself,
- Construct our own index.html that loads P5, then the specified sketch

Closes #4

TODO: Should we auto-open the browser? Should we support Live reload?

Closes #3, because this fixes the issue where useDirectory was only
ever called with process.cwd()

Closes #6 by providing error handling for missing dirs & files